requirements.txt
aiogram==3.10.0
peewee==3.17.6
python-dotenv==1.0.1
phonenumbers==8.13.40
